+++ cpp/inc/bond/core/reflection.h
@@ -93,7 +93,7 @@ struct FieldTemplate
     static const Metadata& metadata;
 
     /// @brief Static data member representing the field pointer
-    static const field_pointer field;
+    static constexpr field_pointer field = field_ptr;
 
     /// @brief Static data member equal to the field ordinal
     BOND_STATIC_CONSTEXPR uint16_t id = field_id;
@@ -125,18 +125,6 @@ template <
     const Metadata* metadata_ptr>
 const bond::Metadata&
     FieldTemplate<field_id, ModifierTag, Struct, FieldType, field_ptr, metadata_ptr>::metadata = *metadata_ptr;
-
-template
-<
-    uint16_t field_id,
-    typename ModifierTag,
-    typename Struct,
-    typename FieldType,
-    FieldType Struct::*field_ptr,
-    const Metadata* metadata_ptr
->
-const typename FieldTemplate<field_id, ModifierTag, Struct, FieldType, field_ptr, metadata_ptr>::field_pointer
-    FieldTemplate<field_id, ModifierTag, Struct, FieldType, field_ptr, metadata_ptr>::field = field_ptr;
 
 
 template <typename Service, typename Input, typename Result> struct